refactor: add schedule maincomponent
