--- src/components/main/ScheduleText/styled.ts
@@ -1,10 +1,0 @@
-import styled from '@emotion/native';
-
-export const HomeScreenScheduleTextContainer = styled.View`
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-  width: auto;
-  flex-direction: column;
-  row-gap: 6px;
-`;
